cleanup: 简化 ClipName/MaterialConstants，删除死代码，小幅度清理
ClipName:
- 从 struct 字符串常量改为 enum(u16)，内建 toString/fromName/toId/fromId
- 删除 clip_name_infos、ClipNameId、clipNameToId/clipNameFromId
- network.zig clip_name_id 从 u8 改为 u16

MaterialConstants:
- 移除 has_base_color/has_normal/_pad8 三个无用字段
- 移除 shader 中的 has_base_color 分支判断，统一 tex × factor × color
- 移除了 setColorTexture/setNormalTexture 中不必要的 syncUniformBuffer

死代码删除:
- systems.zig + systems/health_system.zig（从未被调用）
- block_world.zig: buildMeshSync（被异步版取代）
- block_world.zig: updateAIAgent（被 server.zig updateAIState 取代）

小清理:
- block_world.zig deinit: 移除多余的作用域块
- game.zig: syncCameraFromPlayer 提出重复分支
- shadow.zig: Mat4→[16]f32 手写拷贝改为直接写入（×2 处）

+++ src/network.zig
@@ -30,7 +30,7 @@ pub const EntitySnapshot = struct {
     player_id: u32,
     entity: ECS.Entity,
     entity_type_id: u32,
-    clip_name_id: u8, // 0=idle 1=walk 2=run 3=death 4=attack
+    clip_name_id: u16,
     pos: Vec3,
     facing_yaw: f32,
     facing_pitch: f32,
